add base component
