+++ systemtest/src/main/java/io/strimzi/systemtest/metrics/ClusterOperatorMetricsComponent.java
@@ -5,35 +5,35 @@
 package io.strimzi.systemtest.metrics;
 
 import io.fabric8.kubernetes.api.model.LabelSelector;
-import io.skodjob.testframe.MetricsComponent;
-import io.strimzi.systemtest.TestConstants;
 
 import static io.strimzi.test.k8s.KubeClusterResource.kubeClient;
 
-public class ClusterOperatorMetricsComponent implements MetricsComponent {
+/**
+ * Concrete implementation of BaseMetricsComponent for the Cluster Operator.
+ */
+public class ClusterOperatorMetricsComponent extends BaseMetricsComponent {
 
-    private String namespaceName;
-    private String componentName;
-
+    /**
+     * Factory method to create a new instance of ClusterOperatorMetricsComponent.
+     * @param namespaceName     the namespace in which the component is deployed
+     * @param componentName     the name of the component
+     * @return                  a new instance of ClusterOperatorMetricsComponent
+     */
     public static ClusterOperatorMetricsComponent create(final String namespaceName, final String componentName) {
         return new ClusterOperatorMetricsComponent(namespaceName, componentName);
     }
 
+    /**
+     * Private constructor to enforce the use of the factory method.
+     */
     private ClusterOperatorMetricsComponent(String namespaceName, String componentName) {
-        this.namespaceName = namespaceName;
-        this.componentName = componentName;
+        super(namespaceName, componentName);
     }
 
-    @Override
-    public int getDefaultMetricsPort() {
-        return TestConstants.CLUSTER_OPERATOR_METRICS_PORT;
-    }
-
-    @Override
-    public String getDefaultMetricsPath() {
-        return "/metrics";
-    }
-
+    /**
+     * Provides the label selector specific to the Cluster Operator.
+     * @return LabelSelector for the Cluster Operator deployment
+     */
     @Override
     public LabelSelector getLabelSelector() {
         return kubeClient().getDeploymentSelectors(namespaceName, componentName);
