+++ systemtest/src/main/java/io/strimzi/systemtest/metrics/KafkaBridgeMetricsComponent.java
@@ -5,36 +5,46 @@
 package io.strimzi.systemtest.metrics;
 
 import io.fabric8.kubernetes.api.model.LabelSelector;
-import io.skodjob.testframe.MetricsComponent;
 import io.strimzi.api.kafka.model.bridge.KafkaBridgeResources;
 import io.strimzi.systemtest.TestConstants;
 
 import static io.strimzi.test.k8s.KubeClusterResource.kubeClient;
 
-public class KafkaBridgeMetricsComponent implements MetricsComponent {
+/**
+ * Concrete implementation of BaseMetricsComponent for the Kafka Bridge.
+ */
+public class KafkaBridgeMetricsComponent extends BaseMetricsComponent {
 
-    private String namespaceName;
-    private String componentName;
-
+    /**
+     * Factory method to create a new instance of KafkaBridgeMetricsComponent.
+     * @param namespaceName     the namespace in which the component is deployed
+     * @param componentName     the name of the component
+     * @return                  a new instance of KafkaBridgeMetricsComponent
+     */
     public static KafkaBridgeMetricsComponent create(final String namespaceName, final String componentName) {
         return new KafkaBridgeMetricsComponent(namespaceName, componentName);
     }
 
+    /**
+     * Private constructor to enforce the use of the factory method.
+     */
     private KafkaBridgeMetricsComponent(String namespaceName, String componentName) {
-        this.namespaceName = namespaceName;
-        this.componentName = componentName;
+        super(namespaceName, componentName);
     }
 
+    /**
+     * Provides the default metrics port specifically for the Kafka Bridge.
+     * @return int representing the Kafka Bridge metrics port
+     */
     @Override
     public int getDefaultMetricsPort() {
         return TestConstants.KAFKA_BRIDGE_METRICS_PORT;
     }
 
-    @Override
-    public String getDefaultMetricsPath() {
-        return "/metrics";
-    }
-
+    /**
+     * Provides the label selector specific to the Kafka Bridge.
+     * @return LabelSelector for the Kafka Bridge deployment
+     */
     @Override
     public LabelSelector getLabelSelector() {
         return kubeClient().getDeploymentSelectors(namespaceName, KafkaBridgeResources.componentName(componentName));
