+++ systemtest/src/main/java/io/strimzi/systemtest/metrics/KafkaConnectMetricsComponent.java
@@ -5,33 +5,34 @@
 package io.strimzi.systemtest.metrics;
 
 import io.fabric8.kubernetes.api.model.LabelSelector;
-import io.skodjob.testframe.MetricsComponent;
 import io.strimzi.api.kafka.model.connect.KafkaConnectResources;
-import io.strimzi.systemtest.TestConstants;
 import io.strimzi.systemtest.resources.crd.KafkaConnectResource;
 
-public class KafkaConnectMetricsComponent implements MetricsComponent {
+/**
+ * Concrete implementation of BaseMetricsComponent for Kafka Connect.
+ */
+public class KafkaConnectMetricsComponent extends BaseMetricsComponent {
 
-    private String componentName;
-
+    /**
+     * Factory method to create a new instance of KafkaConnectMetricsComponent.
+     * @param componentName the name of the component
+     * @return a new instance of KafkaConnectMetricsComponent
+     */
     public static KafkaConnectMetricsComponent create(final String componentName) {
         return new KafkaConnectMetricsComponent(componentName);
     }
 
+    /**
+     * Private constructor to enforce the use of the factory method.
+     */
     private KafkaConnectMetricsComponent(String componentName) {
-        this.componentName = componentName;
+        super(null, componentName);
     }
 
-    @Override
-    public int getDefaultMetricsPort() {
-        return TestConstants.COMPONENTS_METRICS_PORT;
-    }
-
-    @Override
-    public String getDefaultMetricsPath() {
-        return "/metrics";
-    }
-
+    /**
+     * Provides the label selector specific to Kafka Connect.
+     * @return LabelSelector for the Kafka Connect deployment
+     */
     @Override
     public LabelSelector getLabelSelector() {
         return KafkaConnectResource.getLabelSelector(componentName, KafkaConnectResources.componentName(componentName));
