+++ systemtest/src/main/java/io/strimzi/systemtest/metrics/KafkaExporterMetricsComponent.java
@@ -5,36 +5,36 @@
 package io.strimzi.systemtest.metrics;
 
 import io.fabric8.kubernetes.api.model.LabelSelector;
-import io.skodjob.testframe.MetricsComponent;
 import io.strimzi.api.kafka.model.kafka.exporter.KafkaExporterResources;
-import io.strimzi.systemtest.TestConstants;
 
 import static io.strimzi.test.k8s.KubeClusterResource.kubeClient;
 
-public class KafkaExporterMetricsComponent implements MetricsComponent {
+/**
+ * Concrete implementation of BaseMetricsComponent for Kafka Exporter.
+ */
+public class KafkaExporterMetricsComponent extends BaseMetricsComponent {
 
-    private String namespaceName;
-    private String componentName;
-
+    /**
+     * Factory method to create a new instance of KafkaExporterMetricsComponent.
+     * @param namespaceName     the namespace in which the component is deployed
+     * @param componentName     the name of the component
+     * @return                  a new instance of KafkaExporterMetricsComponent
+     */
     public static KafkaExporterMetricsComponent create(final String namespaceName, final String componentName) {
         return new KafkaExporterMetricsComponent(namespaceName, componentName);
     }
 
+    /**
+     * Private constructor to enforce the use of the factory method.
+     */
     private KafkaExporterMetricsComponent(String namespaceName, String componentName) {
-        this.namespaceName = namespaceName;
-        this.componentName = componentName;
+        super(namespaceName, componentName);
     }
 
-    @Override
-    public int getDefaultMetricsPort() {
-        return TestConstants.COMPONENTS_METRICS_PORT;
-    }
-
-    @Override
-    public String getDefaultMetricsPath() {
-        return "/metrics";
-    }
-
+    /**
+     * Provides the label selector specific to the Kafka Exporter.
+     * @return LabelSelector for the Kafka Exporter deployment
+     */
     @Override
     public LabelSelector getLabelSelector() {
         return kubeClient().getDeploymentSelectors(namespaceName, KafkaExporterResources.componentName(componentName));
