+++ systemtest/src/main/java/io/strimzi/systemtest/metrics/UserOperatorMetricsComponent.java
@@ -5,35 +5,45 @@
 package io.strimzi.systemtest.metrics;
 
 import io.fabric8.kubernetes.api.model.LabelSelector;
-import io.skodjob.testframe.MetricsComponent;
 import io.strimzi.api.kafka.model.kafka.KafkaResources;
 import io.strimzi.systemtest.TestConstants;
 import io.strimzi.systemtest.resources.ResourceManager;
 
-public class UserOperatorMetricsComponent implements MetricsComponent {
+/**
+ * Concrete implementation of BaseMetricsComponent for the User Operator.
+ */
+public class UserOperatorMetricsComponent extends BaseMetricsComponent {
 
-    private String namespaceName;
-    private String componentName;
-
+    /**
+     * Factory method to create a new instance of UserOperatorMetricsComponent.
+     * @param namespaceName     the namespace in which the component is deployed
+     * @param componentName     the name of the component
+     * @return                  a new instance of UserOperatorMetricsComponent
+     */
     public static UserOperatorMetricsComponent create(final String namespaceName, final String componentName) {
         return new UserOperatorMetricsComponent(namespaceName, componentName);
     }
 
+    /**
+     * Private constructor to enforce the use of the factory method.
+     */
     private UserOperatorMetricsComponent(String namespaceName, String componentName) {
-        this.namespaceName = namespaceName;
-        this.componentName = componentName;
+        super(namespaceName, componentName);
     }
 
+    /**
+     * Provides the default metrics port specifically for the User Operator.
+     * @return int representing the User Operator metrics port
+     */
     @Override
     public int getDefaultMetricsPort() {
         return TestConstants.USER_OPERATOR_METRICS_PORT;
     }
 
-    @Override
-    public String getDefaultMetricsPath() {
-        return "/metrics";
-    }
-
+    /**
+     * Provides the label selector specific to the User Operator.
+     * @return LabelSelector for the User Operator deployment
+     */
     @Override
     public LabelSelector getLabelSelector() {
         return ResourceManager.kubeClient().getDeploymentSelectors(namespaceName, KafkaResources.entityOperatorDeploymentName(componentName));
